+++ app/lib/ui/components/component_list.dart
@@ -1,1 +1,48 @@
+import 'package:flutter/material.dart';
+import 'package:ods_flutter_demo/ui/components/cards/component_card_small.dart';
+import 'package:ods_flutter_demo/ui/components/checkboxes/component_checkboxes.dart';
+import 'package:ods_flutter_demo/ui/components/material/component_material.dart';
+import 'package:ods_flutter_demo/ui/components/radio_buttons/component_radio_buttons.dart';
+import 'package:ods_flutter_demo/ui/components/switches/component_switches.dart';
+import 'package:ods_flutter_demo/ui/components/template/component.dart';
 
+List<OdsComponent> odsComponents = [
+  OdsComponent(
+      "Checkboxes",
+      'assets/ic_checkbox.png',
+      "Checkboxes are used to select or deselect an option. Single checkboxes can be used to determine an individual choice and they can be grouped to allow users to select any combination of items.",
+      [Variant("Checkboxes", "OdsCheckbox", ComponentCheckboxes())]),
+  OdsComponent(
+      "Cards",
+      'assets/ic_card.png',
+      "Cards are important components that can be used to organise and present a number of different types of related information.",
+      [
+        Variant("Small card", "OdsSmallCard", ComponentCardSmall()),
+        Variant("Vertical image first card", "OdsVerticalImageFirstCard",
+            Text("Coming Soon")),
+        Variant("Vertical header first card", "OdsVerticalHeaderFirstCard",
+            Text("Coming Soon")),
+      ]),
+  OdsComponent(
+      "Radio Buttons",
+      'assets/il_radiobutton.png',
+      "Radio buttons enable users to choose one item from a set of mutually exclusive, related choices.",
+      [
+        Variant("Radio Buttons", "OdsRadioButtons", ComponentRadioButtons()),
+      ]),
+  OdsComponent(
+      "Switches",
+      'assets/il_switches.png',
+      "On/off switches toggle the state of a single settings option. The function of the switch should be made clear by the inline label.",
+      [
+        Variant("Switches", "OdsSwitch", ComponentSwitches()),
+      ]),
+  OdsComponent(
+      "Materials",
+      'assets/placeholder.png',
+      "On/off switches toggle the state of a single settings option. The function of the switch should be made clear "
+          "by the inline label.",
+      [
+        Variant("Others", "Material", ComponentMaterial()),
+      ])
+];