+++ app/lib/ui/components/template/components_screen.dart
@@ -2,10 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:ods_flutter/components/card/ods_small_card.dart';
 import 'package:ods_flutter/guidelines/spacings.dart';
-
-import '../../utilities/display_image.dart';
-import 'component.dart';
-import 'component_detail_screen.dart';
+import 'package:ods_flutter_demo/ui/components/component_list.dart';
+import 'package:ods_flutter_demo/ui/components/template/component_detail_screen.dart';
+import 'package:ods_flutter_demo/ui/utilities/display_image.dart';
 
 class ComponentsScreen extends StatefulWidget {
   @override
@@ -13,14 +12,6 @@ class ComponentsScreen extends StatefulWidget {
 }
 
 class _ComponentsScreenState extends State<ComponentsScreen> {
-/*
-class ComponentsScreen extends StatelessWidget {
-  Components group = Components();
-=======
-  Components component = Components();
->>>>>>> a69b6da (Add display image to accept svg file)
-
-*/
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -31,36 +22,15 @@ class ComponentsScreen extends StatelessWidget {
                 child: GridView.count(
                     crossAxisCount: 2,
                     children: odsComponents
-                        .map((c) => OdsSmallCard(
-                              title: c.title,
+                        .map((component) => OdsSmallCard(
+                              title: component.title,
                               image: displayImage(
-                                  c.imageResourceName, double.infinity),
+                                  component.imageResourceName, double.infinity),
                               onTap: () {
-                                Get.to(c.componentScreen);
+                                Get.to(ComponentDetailScreen(
+                                    component: component));
                               },
                             ))
                         .toList()))));
-/*
-                    children:
-                        List.generate(component.getComponent().length, (index) {
-=======
-                    children: List.generate(component.getComponents().length,
-                        (index) {
->>>>>>> 6dbf087 (Add detail page for all components)
-                      return OdsSmallCard(
-                        title: component.getName(index),
-                        image: displayImage(
-                            component.getImage(index), double.infinity),
-                        onTap: () {
-                          Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                  builder: (context) => ComponentDetailScreen(
-                                      component:
-                                          component.getComponent(index))));
-                        },
-                      );
-                    })))));
-*/
   }
 }