+++ app/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Create components list screen ([#132](https://github.com/Orange-OpenSource/ods-flutter/issues/132))
 - Add guideline typography screen ([#121](https://github.com/Orange-OpenSource/ods-flutter/issues/121))
+- Create component description page structure ([#124](https://github.com/Orange-OpenSource/ods-flutter/issues/124))
 
 ### Changed